arch-riscv: Correct vlmax calculation in instructions vlm/vsm
This commit updates the 'micro_vlmax' calculation in
'VlmConstructor' and 'VsmConstructor'.

The output and input vector registers of vlm and vsm
are represented using the mask register layout, which means
a single vector register should be able to represent at most 'vlen' elements.
Therefore, 'lmul' should be considered when calculating 'micro_vlmax'
in the templates.

Change-Id: I97831fcf6246a14211fc0a250d3aa64790cc59f2